// src/app/osm/models/type-category.enum.ts
export enum TypeCategory {
  WASTE_TYPE = 'WASTE_TYPE',
  REGION = 'REGION',
  SUPPLIER_TYPE = 'SUPPLIER_TYPE',
  OLIVE_VARIETY = 'OLIVE_VARIETY',
  OLIVE_TYPE = 'OLIVE_TYPE',
  OPERATION_TYPE = 'OPERATION_TYPE',
  PRODUCTION_METHOD = 'PRODUCTION_METHOD', // Bio / Conventional, etc.
  OIL_VARIETY = 'OIL_VARIETY'
}